requirements.txt
PyQt5==5.15.11
mysql-connector-python==9.5.0
requests==2.32.5
pyinstaller==6.11.1
pytest==8.3.4


